capstone_project/requirements.txt
# Manufacturing Copilot API - Requirements
# Production dependencies for the Manufacturing Copilot system

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries - HuggingFace Ecosystem
transformers==4.35.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4

# LangChain with HuggingFace Integration
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
langgraph==0.0.20

# Vector Database
chromadb==0.4.18

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0

# Monitoring & Logging
prometheus-fastapi-instrumentator==6.1.0

# Note: torch removed - using HuggingFace Inference API instead of local models
# This allows running on CPU-only machines without heavy GPU requirements

# Development & Testing (optional - install with -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
